requirements.txt
akshare>=1.17.85
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
apscheduler>=3.10.0
openpyxl>=3.0.0
pytz>=2023.3
streamlit>=1.28.0
plotly>=5.17.0
supabase>=2.0.0

